cli runs as root user, plugins go missing
Kontena CLI image runs as root user, I think the intent of mounting /home/kontena is to capture .kontena folder with gems etc.  Not sure if this is a change to the image but changing the contena_home volume to mount to at /root seems to fix it.  Thanks for the effort of putting this together, my biggest gripe about kontena is "great first I have to get the correct version of ruby installed..." and you've solved it neatly!